Moved YAML matcher, out of extensions into the core matchers dir
