--- matchers/extensions/yaml/matchers_suite_test.go
@@ -1,13 +1,0 @@
-package yaml_test
-
-import (
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-
-	"testing"
-)
-
-func TestMatchers(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "YAML Matcher Suite")
-}
