+++ matchers/match_yaml_matcher.go
@@ -1,4 +1,4 @@
-package yaml
+package matchers
 
 import (
 	"fmt"
@@ -77,23 +77,4 @@ func (matcher *MatchYAMLMatcher) toStrings(actual interface{}) (actualFormatted,
 	}
 
 	return actualString, expectedString, nil
-}
-
-func toString(a interface{}) (string, bool) {
-	aString, isString := a.(string)
-	if isString {
-		return aString, true
-	}
-
-	aBytes, isBytes := a.([]byte)
-	if isBytes {
-		return string(aBytes), true
-	}
-
-	aStringer, isStringer := a.(fmt.Stringer)
-	if isStringer {
-		return aStringer.String(), true
-	}
-
-	return "", false
 }
